Refactor player card components into separate files
